[MOD] seq에 대해 int -> long으로 수정 [ADD] unit test

+++ sample-springboot-aop/src/main/java/io/simpolor/aop/controller/StudentController.java
@@ -17,7 +17,7 @@ public class StudentController {
 	private StudentService studentService;
 
 	@RequestMapping(value="/totalcount", method=RequestMethod.GET)
-	public int studentTotalCount() {
+	public long studentTotalCount() {
 		return studentService.getStudentTotalCount();
 	}
